refactor: move common utilities to src/common
- Move base_node.py and midi_utils.py to src/common/
- Update import paths in node files
- Update directory structure in documentation

Note: Text formatting issues in documentation will be addressed in a separate PR.
