+++ .clinerules
@@ -13,6 +13,9 @@
 
 - PEP8に従ったコードを書いてください
 - ruffのフォーマッタでファイルの保存と同時に自動整形するので、フォーマットの修正は不要です
+- GoogleスタイルのDocstringを書いてください
+- 設計書の変更が必要あれば設計書を変更してください
+- テストコードを tests ディレクトリ以下に src ディレクトリと同じ構成で作成してください
 
 ## 設計書
