+++ src/common/base_node.py
