+++ src/nodes/0002_synth/synth_node.py
@@ -1,6 +1,6 @@
 import pyxel
-from src.base_node import Node
-from src.midi_utils import MidiNode, SYNTH_PORT, MidiMessage
+from src.common.base_node import Node
+from src.common.midi_utils import MidiNode, SYNTH_PORT, MidiMessage
 
 
 class SynthNode(Node):
